HS-34-03: decompose agent_context.py into an agent_context/ package
The 1,381-line / 46-function module becomes a package with a full re-export
__init__ (63 names), mirroring the Phase-31 db split. No caller or test import
changed.

Layering (acyclic): _common (utils) <- models (constants + AgentSession) <-
hs_context (.hs loader) / hooks (templates + tmux) <- sessions (registry + state
IO + assistant text).

Functions were interleaved by domain, so the split used decorator-aware AST
extraction (start = min decorator lineno; ast.get_source_segment drops
decorators, which would have stripped @dataclass(frozen=True) off AgentSession/
RepoRoot -- a committed test now guards that). Imports trimmed via ruff --fix +
F821 until clean.

Monkeypatch hazards preserved with tests UNCHANGED:
- AGENT_CONTEXT_FILE: readers live in sessions but tests patch the package attr,
  so sessions resolves it through the package (_default_state_file ->
  _agent_context_pkg.AGENT_CONTEXT_FILE) -- the patch reaches the readers.
- shutil.which: __init__ re-exposes `import shutil`; hooks calls shutil.which on
  the same module singleton.
- function patches (get_recent_awaiting_agent_session, ...): auto-preserved by
  the full re-export.

New tests/unit/test_agent_context_package.py guards the public surface, the
frozen dataclasses, and the shutil re-expose. Monkeypatch-sensitive suites (163)
pass; suite green 1958/15; package ruff + F821 clean.
