+++ holdspeak/agent_context/__init__.py
@@ -1,0 +1,152 @@
+"""Agent-session context registry for local coding assistants (HS-34-03 package).
+
+Phase 34 split the 1,381-line agent_context.py module into this package:
+_common (small utils) <- models (constants + AgentSession) <- hs_context
+(.hs loader) / hooks (templates + tmux) <- sessions (registry + state IO).
+The full public surface is re-exported here, so
+`from holdspeak.agent_context import X` is unchanged. shutil is re-exposed so
+tmux tests can monkeypatch holdspeak.agent_context.shutil.which."""
+
+from __future__ import annotations
+
+import shutil  # noqa: F401  (re-exposed for monkeypatch of shutil.which)
+
+from ._common import (
+    _format_timestamp,
+    _optional_str,
+    _parse_timestamp,
+)
+from .models import (
+    AGENT_CONTEXT_FILE,
+    AgentSession,
+    DEFAULT_ASSISTANT_CAPTURE_MAX_CHARS,
+    DEFAULT_CONTEXT_HARD_FILE_MAX_BYTES,
+    DEFAULT_CONTEXT_MAX_BYTES,
+    DEFAULT_CONTEXT_PER_FILE_MAX_BYTES,
+    DEFAULT_PROMPT_CAPTURE_MAX_CHARS,
+    DEFAULT_RECENT_MAX_AGE_SECONDS,
+    DEFAULT_STALE_AGENT_SESSION_SECONDS,
+    HS_CONTEXT_DIR,
+    HS_CONTEXT_FILES,
+    HS_CONTEXT_FILE_KEYS,
+    HS_FLAT_CONTEXT_FILES,
+    HS_IGNORE_FILE,
+    MAX_SESSIONS,
+    STATE_VERSION,
+    SUPPORTED_AGENTS,
+    _SECRET_CONTEXT_RE,
+)
+from .hs_context import (
+    RepoRoot,
+    _derive_project_name,
+    _normalize_project_root,
+    _parse_ignore_lines,
+    _read_context_entry,
+    _read_text_budget,
+    _read_toml_name,
+    compact_hs_project_context,
+    detect_repo_root,
+    load_hs_project_context,
+    render_hs_context_for_prompt,
+)
+from .hooks import (
+    _agent_hook_command,
+    _read_tmux_display,
+    claude_hook_template,
+    codex_hook_template,
+    detect_tmux_context,
+)
+from .sessions import (
+    _bounded_optional_str,
+    _extract_text_from_record,
+    _extract_text_parts,
+    _is_assistant_record,
+    _payload_cwd,
+    _prune_sessions,
+    _read_state,
+    _recent_awaiting_sessions_from_state,
+    _selected_response_key,
+    _session_key,
+    _state_lock,
+    _write_state,
+    clear_agent_session_response,
+    clear_stale_agent_sessions,
+    extract_last_assistant_text,
+    get_recent_agent_session,
+    get_recent_awaiting_agent_session,
+    get_selected_awaiting_agent_session,
+    ingest_agent_hook_event,
+    list_agent_sessions,
+    list_recent_awaiting_agent_sessions,
+    looks_like_agent_question,
+    pin_agent_session,
+    select_awaiting_agent_session,
+    select_next_awaiting_agent_session,
+    set_agent_session_summary,
+)
+
+__all__ = [
+    "AGENT_CONTEXT_FILE",
+    "AgentSession",
+    "DEFAULT_ASSISTANT_CAPTURE_MAX_CHARS",
+    "DEFAULT_CONTEXT_HARD_FILE_MAX_BYTES",
+    "DEFAULT_CONTEXT_MAX_BYTES",
+    "DEFAULT_CONTEXT_PER_FILE_MAX_BYTES",
+    "DEFAULT_PROMPT_CAPTURE_MAX_CHARS",
+    "DEFAULT_RECENT_MAX_AGE_SECONDS",
+    "DEFAULT_STALE_AGENT_SESSION_SECONDS",
+    "HS_CONTEXT_DIR",
+    "HS_CONTEXT_FILES",
+    "HS_CONTEXT_FILE_KEYS",
+    "HS_FLAT_CONTEXT_FILES",
+    "HS_IGNORE_FILE",
+    "MAX_SESSIONS",
+    "RepoRoot",
+    "STATE_VERSION",
+    "SUPPORTED_AGENTS",
+    "_SECRET_CONTEXT_RE",
+    "_agent_hook_command",
+    "_bounded_optional_str",
+    "_derive_project_name",
+    "_extract_text_from_record",
+    "_extract_text_parts",
+    "_format_timestamp",
+    "_is_assistant_record",
+    "_normalize_project_root",
+    "_optional_str",
+    "_parse_ignore_lines",
+    "_parse_timestamp",
+    "_payload_cwd",
+    "_prune_sessions",
+    "_read_context_entry",
+    "_read_state",
+    "_read_text_budget",
+    "_read_tmux_display",
+    "_read_toml_name",
+    "_recent_awaiting_sessions_from_state",
+    "_selected_response_key",
+    "_session_key",
+    "_state_lock",
+    "_write_state",
+    "claude_hook_template",
+    "clear_agent_session_response",
+    "clear_stale_agent_sessions",
+    "codex_hook_template",
+    "compact_hs_project_context",
+    "detect_repo_root",
+    "detect_tmux_context",
+    "extract_last_assistant_text",
+    "get_recent_agent_session",
+    "get_recent_awaiting_agent_session",
+    "get_selected_awaiting_agent_session",
+    "ingest_agent_hook_event",
+    "list_agent_sessions",
+    "list_recent_awaiting_agent_sessions",
+    "load_hs_project_context",
+    "looks_like_agent_question",
+    "pin_agent_session",
+    "render_hs_context_for_prompt",
+    "select_awaiting_agent_session",
+    "select_next_awaiting_agent_session",
+    "set_agent_session_summary",
+]
