+++ pm/roadmap/holdspeak/phase-34-structural-decomposition-ii/current-phase-status.md
@@ -1,9 +1,9 @@
 # Phase 34 — Structural Decomposition II
 
-**Status:** in-progress (opened 2026-06-03). 2/5 stories shipped.
+**Status:** in-progress (opened 2026-06-03). 3/5 stories shipped.
 
-**Last updated:** 2026-06-03 (HS-34-02 shipped — `web/routes/activity.py` → a
-`routes/activity/` sub-package; route table byte-identical, suite green 1956/15).
+**Last updated:** 2026-06-03 (HS-34-03 shipped — `agent_context.py` → an
+`agent_context/` package, full re-export, monkeypatch targets preserved; suite green 1958/15).
 
 ## Goal
 
@@ -53,8 +53,8 @@ modules. No new features; every import path and every HTTP route stays identical
       unchanged. (HS-34-01) ✅
 - [x] `web/routes/activity.py` is a `routes/activity/` sub-package; route table
       identical; `build_activity_router` import unchanged. (HS-34-02) ✅
-- [ ] `agent_context.py` is an `agent_context/` package with a full re-export
-      `__init__`; no caller import changed. (HS-34-03)
+- [x] `agent_context.py` is an `agent_context/` package with a full re-export
+      `__init__`; no caller import changed. (HS-34-03) ✅
 - [ ] `intel.py` is an `intel/` package with a full re-export `__init__`; no caller
       import changed. (HS-34-04)
 - [ ] `uv run pytest -q --ignore=tests/e2e/test_metal.py` green throughout; the
@@ -66,7 +66,7 @@ modules. No new features; every import path and every HTTP route stays identical
 |---|---|---|---|---|
 | HS-34-01 | Split `web/routes/dictation.py` → `routes/dictation/` | done | [story-01-dictation-routes-split.md](./story-01-dictation-routes-split.md) | [evidence-story-01.md](./evidence-story-01.md) |
 | HS-34-02 | Split `web/routes/activity.py` → `routes/activity/` | done | [story-02-activity-routes-split.md](./story-02-activity-routes-split.md) | [evidence-story-02.md](./evidence-story-02.md) |
-| HS-34-03 | Decompose `agent_context.py` → package | not-started | [story-03-agent-context-package.md](./story-03-agent-context-package.md) | — |
+| HS-34-03 | Decompose `agent_context.py` → package | done | [story-03-agent-context-package.md](./story-03-agent-context-package.md) | [evidence-story-03.md](./evidence-story-03.md) |
 | HS-34-04 | Decompose `intel.py` → package | not-started | [story-04-intel-package.md](./story-04-intel-package.md) | — |
 | HS-34-05 | Phase closeout + final-summary | not-started | [story-05-closeout.md](./story-05-closeout.md) | — |
 
@@ -83,8 +83,8 @@ non-hardware-gated.
 
 1. HS-34-01 — `routes/dictation/` split. **✅ done (2026-06-03).**
 2. HS-34-02 — `routes/activity/` split (same pattern; bigger handler count). **✅ done (2026-06-03).**
-3. HS-34-03 — `agent_context/` package (proven Phase-31 re-export pattern). **◀ next.**
-4. HS-34-04 — `intel/` package (same pattern).
+3. HS-34-03 — `agent_context/` package (proven Phase-31 re-export pattern). **✅ done (2026-06-03).**
+4. HS-34-04 — `intel/` package (same pattern). **◀ next.**
 5. HS-34-05 — closeout + final-summary.
 
 The two route splits share a **route-table invariant** check (the app's full route
@@ -107,6 +107,16 @@ the dictation split — every payload shaper is single-group, so no shared state
 thread. Route table byte-identical (38 routes, hash `d4332051064ff059` unchanged);
 committed `test_activity_routes_split.py` locks it. Suite green 1956/15; package
 ruff + F821 clean.
+
+**HS-34-03 shipped** (2026-06-03): `agent_context.py` (1,381L / 46 funcs) → an
+`agent_context/` package (`_common` ← `models` ← `hs_context`/`hooks` ← `sessions`)
+with a full re-export `__init__` (63 names). Done via decorator-aware AST
+extraction (functions were interleaved by domain). The two monkeypatch hazards are
+preserved with tests unchanged: `AGENT_CONTEXT_FILE` is read through the package
+(`_default_state_file()`) so the package-attr patch reaches `sessions`; `shutil` is
+re-exposed for the tmux `shutil.which` patch. Committed
+`test_agent_context_package.py` guards the surface + frozen dataclasses. Suite green
+1958/15; package ruff + F821 clean.
 
 ## Active risks
